clippy sweep, part 1: trivial mechanical lints (~66 warnings fixed)
Part 1 of a multi-commit sweep to clear all clippy warnings on this
branch. Tackles the easy mechanical wins; per-slice commits keep the
diff reviewable.

Lint fixes:
- manual_div_ceil (6): replace `(x + 3) / 4) * 4` with
  `x.next_multiple_of(4)` (ext.rs dir-entry padding) or `.div_ceil()`
  (fat.rs root-dir-sector count, woz_write.rs bit count).
- manual_is_multiple_of (5): `x % 2 != 0` -> `!x.is_multiple_of(2)`
  (hfs_fsck B-tree key alignment, mac_alias path padding,
  chd_options_ui hunk-size validation).
- manual_checked_division (3): `if x > 0 { y / x } else { 0 }` ->
  `y.checked_div(x).unwrap_or(0)` (ext.rs resize reserved-blocks
  scaling, fat.rs cluster-count computation).
- unnecessary_cast (3): drop redundant `u32 as u32` (ext.rs, efs.rs,
  hfsplus.rs).
- clamp (2): `.min(hi).max(lo)` -> `.clamp(lo, hi)` (browse_view tree
  width, dc42 name length).
- manual_iter_find (2): hfs_common bitmap find replaced with
  `(0..max).find(...)`.
- collapsible_match (1): xfs test match on Result -> if let Ok.
- while_let_loop (1): fat.rs FAT chain-walk.
- unused_lifetime (1): hfs_clone find_child<'a>.
- contains_vs_iter_any (1): single_file_chd test.
- unusual_byte_groupings (1): FNV prime hex literal grouped 4-by-4.
- empty_line_after_doc_comment (1): woz.rs.
- ptr_arg (3): drop `&mut Vec<u8>` -> `&mut [u8]` in btree_insert_full
  + repair_extents_btree_structure (neither grows the buffer);
  `&PathBuf` -> `&Path` for the inspect-tab open_browse_zstd helper.
- doc_list_item_without_indentation (21): docstrings used a `+` joiner
  that markdown parsed as a list bullet. Rephrased "X + Y" -> "X plus
  Y" or "X and Y" so the doc renders as prose.
- identity_op / erasing_op (~20): kept the `+ 0x00`, `1 *`, `| 0`
  patterns where they preserve tabular byte-layout readability (e.g.
  `image[off + 0x00..off + 0x04]` mirroring the `+ 0x04` row that
  follows), but wrapped each site in a targeted `#[allow(...)]` with
  a one-line "why kept" comment so future readers know it's
  intentional. One real instance in single_file_chd dropped the
  redundant `1 *`.

cargo build / test (1084 lib + 84 filesystem_e2e) / clippy / fmt
all green. Warnings remaining: 103 (was 169). Next slices:
needless_range_loop, too_many_arguments + type_complexity,
and the long tail.

+++ src/fs/efs.rs
@@ -2362,7 +2362,7 @@ mod tests {
         // later slices will derive the effective location when 0.
         if sb.bmblock != 0 {
             assert!(
-                (sb.bmblock as u32) < sb.fs_size,
+                sb.bmblock < sb.fs_size,
                 "bmblock {} not inside volume of {} blocks",
                 sb.bmblock,
                 sb.fs_size
