+++ src/fs/hfsplus.rs
@@ -4411,8 +4411,8 @@ impl<R: Read + Write + Seek + Send> HfsPlusFilesystem<R> {
     /// Decrement `linkCount` on the inode named `iNode<inode_num>` under the
     /// volume's `HFS+ Private Data` dir. When the count drops to zero, free
     /// the inode's forks, drop its xattrs, and remove the inode catalog row
-    /// + thread + map entry. Bumps `vh.file_count` and the private dir's
-    /// valence accordingly.
+    /// plus thread plus map entry. Bumps `vh.file_count` and the private
+    /// dir's valence accordingly.
     fn delete_hardlink_inode_ref(&mut self, inode_num: u32) -> Result<(), FilesystemError> {
         let private_cnid = self.find_private_dir_cnid()?.ok_or_else(|| {
             FilesystemError::InvalidData(
@@ -5297,6 +5297,7 @@ pub fn validate_hfsplus_integrity(
 }
 
 #[cfg(test)]
+#[allow(clippy::identity_op)] // `1usize * block_size` keeps offset rows aligned
 mod tests {
     use super::*;
 
@@ -5474,7 +5475,7 @@ mod tests {
         let bitmap_block = 1u32;
         let catalog_start_block = 2u32;
         let catalog_blocks = 4u32;
-        let alloc_blocks = 1 + 1 + catalog_blocks as u32; // VH + bitmap + catalog = 6 blocks
+        let alloc_blocks = 1 + 1 + catalog_blocks; // VH + bitmap + catalog = 6 blocks
         let bitmap_data = &mut img[bitmap_block as usize * block_size as usize
             ..(bitmap_block + 1) as usize * block_size as usize];
         // Set bits 0-5 (MSB-first): byte 0 = 0b11111100
